pyproject.toml
[build-system]
requires = ["setuptools>=60", "wheel"]
[project]
name = "uv-wheel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "genson>=1.3.0",
    "ipykernel>=6.25.1",
    "pandas>=1.5.3",
    "pip>=23.2.1",
    "pyyaml>=6.0.0",
    #"deutils @ git+https://github.com/2blo/deutils.git@1.0.0",
]
[dependency-groups]
# dependency-groups = {deutils = {file = ["deutils.requirements.txt"]}}
#deutils = ["deutils @ git+https://github.com/2blo/deutils.git@1.0.0"]
dev = ["pip>=23.2.1", "uv-wheel[deutils,databricks-15-4]"]
dynamic = ["optional-dependencies"]
[tools.setuptools.dynamic]
optional-dependencies = { deutils = { file = [
    "deutils.requirements.txt",
] }, databricks-15-4 = { file = [
    "databricks.15.4.requirements.txt",
] } }
